tempArticle persistence adapter
