+++ system/blog-core/adapter/acl-user/src/main/kotlin/kr/co/jiniaslog/blogcore/adapter/acl/user/UserAcl.kt
@@ -1,7 +1,7 @@
 package kr.co.jiniaslog.blogcore.adapter.acl.user
 
-import kr.co.jiniaslog.blogcore.application.article.infra.UserServiceClient
 import kr.co.jiniaslog.blogcore.domain.article.UserId
+import kr.co.jiniaslog.blogcore.domain.article.UserServiceClient
 import kr.co.jiniaslog.shared.core.context.AntiCorruptLayer
 import kr.co.jiniaslog.user.application.UserServiceStub
