+++ system/blog-core/adapter/http-api/src/main/kotlin/kr/co/jiniaslog/blogcore/adapter/http/Reqeuests.kt
@@ -1,27 +1,27 @@
 package kr.co.jiniaslog.blogcore.adapter.http
 
 import kr.co.jiniaslog.blogcore.application.article.usecase.ArticleEditCommand
-import kr.co.jiniaslog.blogcore.application.article.usecase.TempArticlePostCommand
+import kr.co.jiniaslog.blogcore.application.article.usecase.DraftArticlePostCommand
 import kr.co.jiniaslog.blogcore.domain.article.ArticleId
 import kr.co.jiniaslog.blogcore.domain.article.UserId
 import kr.co.jiniaslog.blogcore.domain.category.CategoryId
 import kr.co.jiniaslog.blogcore.domain.tag.TagId
 
-data class ArticlePostRequest(
+data class DraftArticlePostRequest(
     val writerId: Long,
     val title: String,
     val content: String,
-    val thumbnailUrl: String,
-    val categoryId: Long,
+    val thumbnailUrl: String?,
+    val categoryId: Long?,
     val tags: Set<Long>,
 ) {
-    fun toCommand(): TempArticlePostCommand {
-        return TempArticlePostCommand(
+    fun toCommand(): DraftArticlePostCommand {
+        return DraftArticlePostCommand(
             userId = UserId(writerId),
             title = title,
             content = content,
             thumbnailUrl = thumbnailUrl,
-            categoryId = CategoryId(categoryId),
+            categoryId = if (categoryId == null) null else CategoryId(categoryId),
             tags = tags.map { TagId(it) }.toSet(),
         )
     }
