--- system/blog-core/adapter/persistence/article-rds/src/main/resources/article_persistence-application.yml
@@ -1,17 +1,0 @@
-spring:
-  #===========#
-  #   MySQL   #
-  #===========#
-  datasource:
-    article:
-      jdbc-url: jdbc:mysql://localhost:3307/jinia_blog_article?serverTimezone=Asia/Seoul
-      username: root
-      password: 1
-      driver-class-name: com.mysql.cj.jdbc.Driver
-
-  flyway:
-    locations: db.migration.article
-    enabled: true
-    baseline-on-migrate: true          # flyway_schema_history 테이블을 자동으로 생성할지 여부
-    baseline-version: 0
-    validateMigrationNaming: true
