+++ system/blog-core/adapter/persistence/core-rds/src/main/kotlin/kr/co/jiniaslog/blogcore/adapter/persistence/article/ArticlePM.kt
