+++ system/blog-core/adapter/persistence/core-rds/src/main/kotlin/kr/co/jiniaslog/blogcore/adapter/persistence/article/JpaTempArticleRepository.kt
@@ -1,0 +1,5 @@
+package kr.co.jiniaslog.blogcore.adapter.persistence.article
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface JpaTempArticleRepository : JpaRepository<TempArticlePM, Long>
