+++ system/blog-core/adapter/persistence/core-rds/src/main/kotlin/kr/co/jiniaslog/blogcore/adapter/persistence/article/TempArticlePmMapper.kt
@@ -1,0 +1,19 @@
+package kr.co.jiniaslog.blogcore.adapter.persistence.article
+
+import kr.co.jiniaslog.blogcore.domain.article.TempArticle
+import org.mapstruct.InjectionStrategy
+import org.mapstruct.Mapper
+import org.mapstruct.Mapping
+import org.mapstruct.MappingConstants
+import org.mapstruct.ReportingPolicy
+
+@Mapper(
+    componentModel = MappingConstants.ComponentModel.SPRING,
+    injectionStrategy = InjectionStrategy.CONSTRUCTOR,
+    unmappedTargetPolicy = ReportingPolicy.ERROR,
+)
+interface TempArticlePmMapper {
+    @Mapping(source = "id.value", target = "id")
+    @Mapping(source = "writerId.value", target = "writerId")
+    fun toPm(article: TempArticle): TempArticlePM
+}
