+++ system/blog-core/adapter/persistence/core-rds/src/main/resources/blogcore_persistence-application.yml
@@ -1,0 +1,17 @@
+spring:
+  #===========#
+  #   MySQL   #
+  #===========#
+  datasource:
+    blogcore:
+      jdbc-url: jdbc:mysql://localhost:3307/jinia_blog_core?serverTimezone=Asia/Seoul
+      username: root
+      password: 1
+      driver-class-name: com.mysql.cj.jdbc.Driver
+
+  flyway:
+    locations: db.migration.blogcore
+    enabled: true
+    baseline-on-migrate: true          # flyway_schema_history 테이블을 자동으로 생성할지 여부
+    baseline-version: 0
+    validateMigrationNaming: true
