+++ system/blog-core/adapter/persistence/core-rds/src/main/resources/db.migration.blogcore/V1__create_article_tbl.sql
