+++ system/blog-core/adapter/persistence/core-rds/src/main/resources/db.migration.blogcore/V2__create_temp_article_tbl.sql
@@ -1,0 +1,11 @@
+CREATE TABLE temp_article
+(
+    temp_article_id    BIGINT         NOT NULL,
+    title         VARCHAR(50)    NULL,
+    content       VARCHAR(10000) NULL,
+    thumbnail_url VARCHAR(255)   NULL,
+    writer_id     BIGINT         NOT NULL,
+    created_date  DATETIME(6),
+    updated_date  DATETIME(6),
+    PRIMARY KEY (temp_article_id)
+);
