--- system/blog-core/application/src/main/kotlin/kr/co/jiniaslog/blogcore/application/article/infra/ArticleIdGenerator.kt
@@ -1,7 +1,0 @@
-package kr.co.jiniaslog.blogcore.application.article.infra
-
-import kr.co.jiniaslog.blogcore.domain.article.ArticleId
-
-interface ArticleIdGenerator {
-    fun generate(): ArticleId
-}
