--- system/blog-core/application/src/main/kotlin/kr/co/jiniaslog/blogcore/application/article/infra/UserServiceClient.kt
@@ -1,5 +1,0 @@
-package kr.co.jiniaslog.blogcore.application.article.infra
-
-interface UserServiceClient {
-    fun isAdmin(id: Long): Boolean
-}
