+++ system/blog-core/application/src/main/kotlin/kr/co/jiniaslog/blogcore/application/article/usecase/TempArticleUseCases.kt
@@ -19,3 +19,5 @@ data class TempArticlePostCommand(
     val categoryId: CategoryId?,
     val tags: Set<TagId>,
 )
+
+data class TempArticlePostCommandInValidException(override val message: String) : RuntimeException(message)
