+++ system/blog-core/application/src/main/kotlin/kr/co/jiniaslog/blogcore/application/article/usecase/TempArticleUseCasesInteractor.kt
@@ -1,20 +1,21 @@
 package kr.co.jiniaslog.blogcore.application.article.usecase
 
 import kr.co.jiniaslog.blogcore.application.article.infra.TransactionHandler
-import kr.co.jiniaslog.blogcore.application.article.infra.UserServiceClient
+import kr.co.jiniaslog.blogcore.domain.article.ArticleId
 import kr.co.jiniaslog.blogcore.domain.article.TempArticle
-import kr.co.jiniaslog.blogcore.domain.article.TempArticleService
+import kr.co.jiniaslog.blogcore.domain.article.TempArticleRepository
+import kr.co.jiniaslog.blogcore.domain.article.UserServiceClient
 import kr.co.jiniaslog.shared.core.context.UseCaseInteractor
 
 @UseCaseInteractor
 internal class TempArticleUseCasesInteractor(
     private val transactionHandler: TransactionHandler,
-    private val tempArticleService: TempArticleService,
+    private val TempArticleRepository: TempArticleRepository,
     private val userServiceClient: UserServiceClient,
 ) : TempArticleUseCases {
 
     override fun post(command: TempArticlePostCommand) = with(command) {
-        if (this.isInvalid()) throw IllegalArgumentException("Invalid ArticlePostCommand")
+        if (command.isInvalid()) throw TempArticlePostCommandInValidException("Invalid ArticlePostCommand : $command")
 
         val newTemp = TempArticle.Factory.newTempOne(
             userId = userId,
@@ -25,16 +26,16 @@ internal class TempArticleUseCasesInteractor(
             tags = tags,
         )
 
-        transactionHandler.runInReadCommittedTransaction { tempArticleService.saveTempArticle(newTemp) }
+        transactionHandler.runInReadCommittedTransaction { TempArticleRepository.save(newTemp) }
     }
 
     private fun TempArticlePostCommand.isInvalid(): Boolean =
         !userServiceClient.isAdmin(userId.value)
 
     override fun delete() {
-        transactionHandler.runInReadCommittedTransaction { tempArticleService.deleteTempArticle() }
+        transactionHandler.runInReadCommittedTransaction { TempArticleRepository.delete() }
     }
 
     override fun findOne(): TempArticle? =
-        tempArticleService.findTempArticle()
+        TempArticleRepository.findTemp(ArticleId(TempArticle.TEMP_ARTICLE_STATIC_ID))
 }
