+++ system/blog-core/domain/src/main/kotlin/kr/co/jiniaslog/blogcore/domain/article/ArticleIdGenerator.kt
@@ -1,0 +1,5 @@
+package kr.co.jiniaslog.blogcore.domain.article
+
+interface ArticleIdGenerator {
+    fun generate(): ArticleId
+}
