+++ system/blog-core/domain/src/main/kotlin/kr/co/jiniaslog/blogcore/domain/article/ArticleRepository.kt
@@ -1,0 +1,7 @@
+package kr.co.jiniaslog.blogcore.domain.article
+
+interface ArticleRepository {
+    fun save(newArticle: Article)
+
+    fun findById(articleId: ArticleId): Article?
+}
