+++ system/blog-core/domain/src/main/kotlin/kr/co/jiniaslog/blogcore/domain/article/UserServiceClient.kt
@@ -1,0 +1,5 @@
+package kr.co.jiniaslog.blogcore.domain.article
+
+interface UserServiceClient {
+    fun isAdmin(id: Long): Boolean
+}
