+++ system/infra/src/main/kotlin/kr/co/jiniaslog/infra/events/EventPublishExecutorConfig.kt
@@ -1,0 +1,21 @@
+package kr.co.jiniaslog.infra.events
+
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor
+
+@Configuration
+internal class EventPublishExecutorConfig {
+    @Bean
+    fun domainEventPublishExecutor(): ThreadPoolTaskExecutor {
+        val executor = ThreadPoolTaskExecutor()
+        executor.corePoolSize = 10
+        executor.queueCapacity = 100
+        executor.setAllowCoreThreadTimeOut(true)
+        executor.setThreadNamePrefix("domainEventPublishExecutor-")
+        executor.setWaitForTasksToCompleteOnShutdown(true)
+        executor.setAwaitTerminationSeconds(300)
+        executor.initialize()
+        return executor
+    }
+}
